Changed a little bit on the tenant definition and added the crud service for tenants.
